Implement deserialize File type object method

+++ src/main/java/pl/jkuznik/computer/software/file/musicfile/MP3MusicFile.java
@@ -5,7 +5,7 @@ import pl.jkuznik.computer.software.file.FileType;
 public class MP3MusicFile extends AbstractMusicFile {
     private int quality;
 
-    public MP3MusicFile(String name, int size, String bandName, String title, int quality) {
+    public MP3MusicFile(String name, long size, String bandName, String title, int quality) {
         super(name, size, bandName, title);
 
         this.quality = quality;
